Remove unused build-depends
p6doc and Pod::To::Markdown are not used at all. META6 is used, but as an authoring dependency ( not the same as a build dependency ) for which there is no documented field ( yet ). Without this change anytime someone tries to install this module it will try to install META6 p6doc and Pod::To::Markdown despite them never being needed.